requirements.txt
asgiref==3.2.5
beautifulsoup4==4.9.1
certifi==2020.4.5.1
Django==3.0.3
django-htmlmin==0.11.0
feedparser==5.2.1
html5lib==1.0.1
psycopg2==2.8.5
pytz==2019.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.0
webencodings==0.5.1
wincertstore==0.2
